Move models to models folder and update imports
